requirements.txt
dvc
dvc[gdrive]
sklearn
pyyaml
pandas